replaced routes to separate file. otherwise the error pops when making browser and memory router in the same file. Tests passing. Memory Router doesnt load Home page right away, but it is working
