+++ src/routes/routes.jsx
@@ -1,0 +1,64 @@
+import App from "../App.jsx";
+import { Home, loader as homeLoader } from "../Pages/Home/Home";
+import { Auth_Layout } from "../Pages/Auth/Auth_Layout";
+import Login from "../Pages/Auth/Login";
+import Register from "../Pages/Auth/Register";
+import { Error } from "../Pages/Error/Error";
+import { Pages, loader as pagesLoader } from "../Pages/Pages/Pages.jsx";
+import Test from "../Pages/Test/Test.jsx";
+import Cards from "../Pages/Services/Cards.jsx";
+
+const routes = [
+	{
+		path: "/",
+		element: <App />,
+		errorElement: <Error />,
+		children: [
+			{
+				element: <Home />,
+				index: true,
+				loader: homeLoader,
+			},
+			/*
+			it should be this:
+
+			{
+				path: "auth",
+				element: <Auth_Layout />,
+				children: [
+					{ path: "login", element: <Login /> },
+					{ path: "register", element: <Register /> },
+				],
+			},
+
+			*/
+			{
+				path: "auth",
+				element: <Auth_Layout />,
+			},
+			{
+				path: "auth/login",
+				element: <Login />,
+			},
+			{
+				path: "auth/register",
+				element: <Register />,
+			},
+			{
+				path: "pages/:pageName",
+				element: <Pages />,
+				loader: pagesLoader,
+			},
+			{
+				path: "pages",
+				element: <Cards />,
+			},
+			{
+				path: "test",
+				element: <Test />,
+			},
+		],
+	},
+];
+
+export { routes };
